wellbeing_project/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
cobble==0.1.3
Django==2.2.4
django-crispy-forms==1.13.0
django-filter==21.1
django-hosts==3.0
django-tenant-schemas==1.9.0
django-tenants==2.2.3
django-utils-six==2.0
djangorestframework==3.12.4
idna==3.3
mammoth==1.4.17
numpy==1.21.2
ordered-set==4.0.2
Pillow==8.3.2
psycopg2==2.9.1
psycopg2-binary==2.8.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
